requirements.txt
# ウェブフレームワーク
fastapi>=0.95.0
uvicorn>=0.21.1
streamlit>=1.22.0

# HTTP/APIクライアント
requests>=2.28.2
httpx>=0.24.0

# データ処理
pydantic>=1.10.7
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
tomli>=2.0.1

# Reddit API
praw>=7.7.0

# arXiv API
arxiv>=1.4.7

# OpenAI API互換クライアント
openai>=0.27.0

# ユーティリティ
tqdm>=4.65.0
tenacity>=8.2.2

# 日付処理
python-dateutil>=2.8.2

